removing redundant code
