+++ helpers/getTrimmedEvents.ts
@@ -1,0 +1,20 @@
+export const getTrimmedEvent = (events) =>
+  events?.pages.flat().flatMap((item) =>
+    item.eventsCalendarConnection.edges.map((edge) => ({
+      ...edge.node,
+      startDateTime: new Date(edge.node.startDateTime),
+      endDateTime: new Date(edge.node.endDateTime),
+      category: formatCategory(edge.node.category),
+    }))
+  ) || [];
+
+export const formatCategory = (category: string): string => {
+  {
+    const categoryReplacements = {
+      "Non-English Courses": "Other",
+    };
+    const lookup = categoryReplacements[category];
+
+    return lookup ? lookup : category;
+  }
+};
