+++ hooks/useFetchEvents.ts
@@ -1,4 +1,5 @@
 import { EventFilterAllCategories } from "@/components/filter/FilterBlock";
+import { formatCategory, getTrimmedEvent } from "@/helpers/getTrimmedEvents";
 import { EVENTS_MAX_SIZE_OVERRIDE } from "@/services/server/getEvents";
 import { GetPastEventsQueryQuery } from "@/tina/types";
 import { useInfiniteQuery } from "@tanstack/react-query";
@@ -18,17 +19,6 @@ const getCategoriesForFilter = (category: string) => {
   const lookup = categories[category];
 
   return lookup ? lookup : [category];
-};
-
-const formatCategory = (category: string): string => {
-  {
-    const categoryReplacements = {
-      "Non-English Courses": "Other",
-    };
-    const lookup = categoryReplacements[category];
-
-    return lookup ? lookup : category;
-  }
 };
 
 export const TODAY = new Date();
@@ -69,15 +59,7 @@ export const useFetchFutureEvents = (filters: SelectedCategories) => {
       },
     });
   return {
-    futureEvents:
-      data?.pages.flat().flatMap((item) =>
-        item.eventsCalendarConnection.edges.map((edge) => ({
-          ...edge.node,
-          startDateTime: new Date(edge.node.startDateTime),
-          endDateTime: new Date(edge.node.endDateTime),
-          category: formatCategory(edge.node.category),
-        }))
-      ) || [],
+    futureEvents: getTrimmedEvent(data),
     error,
     isLoadingFuturePages: isLoading,
     fetchFutureNextPage: fetchNextPage,
@@ -117,15 +99,7 @@ export const useFetchPastEvents = (filters: SelectedCategories) => {
     });
 
   return {
-    pastEvents:
-      data?.pages.flat().flatMap((item) =>
-        item.eventsCalendarConnection.edges.map((edge) => ({
-          ...edge.node,
-          startDateTime: new Date(edge.node.startDateTime),
-          endDateTime: new Date(edge.node.endDateTime),
-          category: formatCategory(edge.node.category),
-        }))
-      ) || [],
+    pastEvents: getTrimmedEvent(data),
     error,
     isLoadingPastPages: isLoading,
     fetchNextPastPage: fetchNextPage,
